pyproject.toml
[project]
name = "robo"
version = "0.1.0"
description = "Discord bot test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
]
